Add files via upload
These files are used for proof of concept and require that the user gain an access token for the Spotify Web Player API and paste that token into the initialize.js.  The user must then select the Spotify Web App as the player through the Spotify remote control feature. The player.html web app with then begin playback and display the next two songs in the queue. The player UI will also display and track a songs playback.